+++ krita/ui/kis_zoom_manager.h
@@ -55,7 +55,7 @@ public:
 
 private slots:
 
-    void slotZoomChanged(KoZoomMode::Mode mode, int zoom);
+    void slotZoomChanged(KoZoomMode::Mode mode, double zoom);
     void slotActualSize();
     void toggleShowRulers(bool show);
     void mousePositionChanged(const QPoint &pos);
